+++ src/dictionarly/hooks/index.ts
@@ -1,3 +1,4 @@
 export { useDictionarlyPlayProgress } from './use-dictionarly-play-progress.hook'
 export { useDictionarlyPlay } from './use-dictionarly-play.hook'
 export { usePlayDictionarly } from './use-play-dictionarly.hook'
+export { useDictionarlySpy } from './use-dictionarly-spy.hook'